mean_reversion/rsi2_spy_tlt_strategy/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pandas_ta>=0.3.14b
yfinance>=0.2.0
pyyaml>=6.0
matplotlib>=3.5.0
